refactor: break down SubmitTx into separate files
