+++ .papi/descriptors/package.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.1.0-autogenerated.8494276640227355333",
+  "version": "0.1.0-autogenerated.3203825568190824154",
   "name": "@polkadot-api/descriptors",
   "files": [
     "dist"
